libs/trader/pyproject.toml
[project]
name = "kamas-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs>=25.4.0",
    "numpy>=2.3.5",
]
